Refactor

+++ tests/api/utils/userSession.ts
@@ -1,58 +1,69 @@
 import { request, APIRequestContext, expect } from '@playwright/test';
-import { ApiHelper } from './apiHelper';
+import { Env } from '../../../config/env';
 
 export class UserSession {
   private api!: APIRequestContext;
-  private helper: ApiHelper;
   userId: string | null = null;
   token: string | null = null;
-
-  constructor(helper: ApiHelper) {
-    this.helper = helper;
-  }
+  cleanUp: boolean | null = true;
 
   async init(): Promise<void> {
-    const env = this.helper.getEnv();
     this.api = await request.newContext({
-      baseURL: env.baseUrl,
-      extraHTTPHeaders: env.headers
+      baseURL: Env.API_URL,
+      extraHTTPHeaders: Env.API_HEADERS,
     });
   }
 
-  async createUser(userName: string, password: string): Promise<string> {
+  async createUser(userName: string, password: string, cleanUp: boolean): Promise<string> {
+    this.cleanUp = cleanUp;
     const response = await this.api.post('/Account/v1/User', {
-      data: { userName, password }
+      data: { userName, password },
     });
-    const body = await response.json();
-    expect(response.status()).toBe(201);
-    this.userId = body.userID || body.userId;
 
-    // genera token automáticamente
-    await this.generateToken(userName, password);
+    if (response.status() === 201) {
+      const body = await response.json();
+      this.userId = body.userID || body.userId;
+      await this.generateToken(userName, password);
+      return this.userId!;
+    }
 
-    return this.userId!;
+    throw new Error(`Unexpected error in createUser(): ${response.status()}`);
   }
 
   async generateToken(userName: string, password: string): Promise<string> {
     const response = await this.api.post('/Account/v1/GenerateToken', {
-      data: { userName, password }
+      data: { userName, password },
     });
-    const body = await response.json();
+
+    const body = await response.json().catch(() => {
+      throw new Error('Error parsing token response');
+    });
+
     expect(response.status()).toBe(200);
     this.token = body.token;
-    console.log(`Token generated: ${this.token?.slice(0, 15)}...`);
+    console.log(`Token generated.`);
     return this.token!;
   }
 
   async deleteUser(): Promise<void> {
+    console.warn("ENTRO DELETE??");
+    console.warn(`this user ID: ${this.userId}`);
+    console.warn(`this token: ${this.token}`);
+    console.warn(`this CLEAN UP: ${this.cleanUp}`);
+    if (!this.cleanUp) {
+      console.warn('No Cleaning data')
+      return;
+    }
     if (!this.userId || !this.token) {
       console.log('No user or token to delete.');
       return;
     }
-    const response = await this.api.delete(`/Account/v1/User/${this.userId}`, {
-      headers: { Authorization: `Bearer ${this.token}` }
+
+    await this.api.delete(`/Account/v1/User/${this.userId}`, {
+      headers: { Authorization: `Bearer ${this.token}` },
     });
-    console.log(`User ${this.userId} deleted → status ${response.status()}`);
+
+    console.log(`User ${this.userId} deleted.`);
   }
 
   async dispose(): Promise<void> {
